// src/app/api/model/apiUserCreatePostRequest.ts
/**
 * ChatWave API Documentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiUserCreatePostRequest { 
    name?: string;
    username?: string;
    email?: string;
    phone?: string;
    password?: string;
    gender?: ApiUserCreatePostRequest.GenderEnum;
}
export namespace ApiUserCreatePostRequest {
    export const GenderEnum = {
        Male: 'male',
        Female: 'female',
        Other: 'other'
    } as const;
    export type GenderEnum = typeof GenderEnum[keyof typeof GenderEnum];
}